+++ models/ConformanceResourceTableData.ts
@@ -1,10 +1,12 @@
+import { FhirpathLabUseContexts } from "~/helpers/useContext_helpers";
 import { BaseResourceData, BaseResourceTableDefinition } from "./BaseResourceTableData";
 
 export interface ConformanceResourceTableDefinition<TData> extends BaseResourceTableDefinition<TData> {
   searchForStatus?: string;
   searchForUseContext?: { system?: string, code: string; display: string; }[];
 
-  searchUseContexts?: { system?: string, code: string; display: string; }[];
+  searchUseContexts?: FhirpathLabUseContexts[];
+  defaultUseContexts?: FhirpathLabUseContexts[];
   searchPublishingStatuses: string[];
 }